+++ tests/test_mpo.py
@@ -15,33 +15,6 @@ from tenpy.networks.terms import OnsiteTerms, CouplingTerms, MultiCouplingTerms,
 spin_half = site.SpinHalfSite(conserve='Sz')
 
 
-def check_hermitian(H):
-    """Check if `H` is a hermitian MPO."""
-    if not H.finite:
-        # include once over the boundary: double the unit cell
-        # a general MPO might have terms going over multiple unit cells, but we ignore that...
-        Ws = H._W * 2
-    else:
-        Ws = H._W
-    #check trace(H.H) = trace(H.H^dagger):  equivalent to H=H^dagger
-    W = Ws[0].take_slice([H.get_IdL(0)], ['wL'])
-
-    trHH = npc.tensordot(W, W.replace_label('wR', 'wR*'), axes=[['p', 'p*'], ['p*', 'p']])
-    trHHd = npc.tensordot(W, W.conj(), axes=[['p', 'p*'], ['p*', 'p']])
-    for W in Ws[1:]:
-        trHH = npc.tensordot(trHH, W, axes=['wR', 'wL'])
-        trHHd = npc.tensordot(trHHd, W, axes=['wR', 'wL'])
-        trHH = npc.tensordot(trHH,
-                             W.replace_label('wR', 'wR*'),
-                             axes=[['wR*', 'p', 'p*'], ['wL', 'p*', 'p']])
-        trHHd = npc.tensordot(trHHd, W.conj(), axes=[['wR*', 'p', 'p*'], ['wL*', 'p*', 'p']])
-    i = H.get_IdR(H.L - 1)
-    trHH = trHH[i, i]
-    trHHd = trHHd[i, i]
-    print("check_hermitian: ", trHH, trHHd)
-    npt.assert_array_almost_equal_nulp(trHH, trHHd, H.L * 20)
-
-
 def test_MPO():
     s = spin_half
     for bc in mpo.MPO._valid_bc:
@@ -60,7 +33,7 @@ def test_MPO():
             H.test_sanity()
             print(H.dim)
             print(H.chi)
-            check_hermitian(H)
+            assert H.is_hermitian()
         if L == 4:
             H2 = H.group_sites(n=2)
             H2.test_sanity()
@@ -198,6 +171,26 @@ def test_MPOEnvironment():
         if E_old is not None:
             assert (abs(E - E_old) < 1.e-14)
         E_old = E
+
+
+def test_MPO_hermitian():
+    s = spin_half
+    ot = OnsiteTerms(4)
+    ct = CouplingTerms(4)
+    ct.add_coupling_term(1., 2, 3, 'Sm', 'Sp')
+    H = mpo.MPOGraph.from_terms(ot, ct, [s] * 4, 'infinite').build_MPO()
+    #  assert not H.is_hermitian()
+    ct.add_coupling_term(1., 2, 3, 'Sp', 'Sm')
+    H = mpo.MPOGraph.from_terms(ot, ct, [s] * 4, 'infinite').build_MPO()
+    #  assert H.is_hermitian()
+
+    ct.add_coupling_term(1., 3, 18, 'Sm', 'Sp')
+    H = mpo.MPOGraph.from_terms(ot, ct, [s] * 4, 'infinite').build_MPO()
+    assert not H.is_hermitian()
+
+    ct.add_coupling_term(1., 3, 18, 'Sp', 'Sm')
+    H = mpo.MPOGraph.from_terms(ot, ct, [s] * 4, 'infinite').build_MPO()
+    assert H.is_hermitian()
 
 
 def test_MPO_expectation_value():