+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Changed
 Added
 ^^^^^
 - `max_range` attribute in :class:`~tenpy.networks.mpo.MPO` and :class:`~tenpy.networks.mpo.MPOGraph`.
+- :meth:`~tenpy.networks.mpo.MPO.is_hermitian`
 
 
 [0.4.0] - 2019-04-28